manual update, added requirements file
In rare occasions, python modules can clash, resulting in AIM not being able to run. While this is not caused by AIM, we'd still like to make sure everyone can use AIM, and solve this kind of issue. To do this, a requirements file has been added, to help avoid/solve these issues.

The manual has been updated to include explanation on the use of this file. Some other updates have been made to the manual, too.

Updated the readme to include an example of running AIM in parallel.